### 2023-05-09 #5455 @Kevin-Mattheus-Moerman
@CheWeiChang1992 thanks for this submission. I am the AEiC on this track and here to help process initial steps. For the moment can you try to address the following:

- [ ] Please fix the potentially missing/invalid DOIs pointed out above :point_up: , you can call `@editorialbot check references` to check them again, and use `@editorialbot generate pdf` to update the paper.
- [ ] Our system did not detect an open source license. Please can you clarify what license your software uses? Also I recommend you add a plain text LICENSE file in the main directory of your repository.
