### 2023-05-10 1e061e7
fix DOIs issue detected by JOSS editorialbot
