+++ paper/paper.bib
@@ -29,7 +29,7 @@
   volume={9},
   pages={e54507},
   year={2020},
-  DOI={https://doi.org/10.7554/eLife.54507},
+  DOI={10.7554/eLife.54507},
   publisher={eLife Sciences Publications, Ltd}
 }
 
@@ -41,7 +41,7 @@
   number={7},
   pages={1106--1108},
   year={2016},
-  DOI={https://doi.org/10.1093/bioinformatics/btv703},
+  DOI={10.1093/bioinformatics/btv703},
   publisher={Oxford University Press}
 }
 @article{gutaker2020genomic,
@@ -52,7 +52,7 @@
   number={5},
   pages={492--502},
   year={2020},
-  DOI={https://doi.org/10.1038/s41477-020-0659-6},
+  DOI={10.1038/s41477-020-0659-6},
   publisher={Nature Publishing Group UK London}
 }
 @article{house2018evaluating,
@@ -63,7 +63,7 @@
   number={3},
   pages={448--460},
   year={2018},
-  DOI={https://doi.org/10.1111/1755-0998.12747},
+  DOI={10.1111/1755-0998.12747},
   publisher={Wiley Online Library}
 }
 @article{konig2020bridge,
@@ -73,7 +73,7 @@
   volume={11},
   pages={701},
   year={2020},
-  DOI={https://doi.org/10.3389/fpls.2020.00701},
+  DOI={10.3389/fpls.2020.00701},
   publisher={Frontiers Media SA}
 }
 @article{lasky2023genotype,
@@ -84,7 +84,7 @@
   number={1},
   pages={125--138},
   year={2023},
-  DOI={https://doi.org/10.1093/plcell/koac267},
+  DOI={10.1093/plcell/koac267},
   publisher={Oxford University Press}
 }
 @article{li2019identifying,
@@ -95,7 +95,7 @@
   number={15},
   pages={3544--3560},
   year={2019},
-  DOI={https://doi.org/10.1111/mec.15169},
+  DOI={10.1111/mec.15169},
   publisher={Wiley Online Library}
 }
 @article{liu2020pan,
@@ -106,7 +106,7 @@
   number={1},
   pages={162--176},
   year={2020},
-  DOI={https://doi.org/10.1016/j.cell.2020.05.023},
+  DOI={10.1016/j.cell.2020.05.023},
   publisher={Elsevier}
 }
 
@@ -118,7 +118,7 @@
   number={2},
   pages={319--326},
   year={2019},
-  DOI={https://doi.org/10.1038/s41588-018-0266-x},
+  DOI={10.1038/s41588-018-0266-x},
   publisher={Nature Publishing Group US New York}
 }
 @article{schulthess2022genomics,
@@ -129,7 +129,7 @@
   number={10},
   pages={1544--1552},
   year={2022},
-  DOI={https://doi.org/10.1038/s41588-022-01189-7},
+  DOI={10.1038/s41588-022-01189-7},
   publisher={Nature Publishing Group US New York}
 }
 @article{wang2018genomic,
@@ -140,7 +140,7 @@
   number={7703},
   pages={43--49},
   year={2018},
-  DOI={https://doi.org/10.1038/s41586-018-0063-9},
+  DOI={10.1038/s41586-018-0063-9},
   publisher={Nature Publishing Group UK London}
 }
 @article{yang2012model,
@@ -151,6 +151,6 @@
   number={6},
   pages={725--731},
   year={2012},
-  DOI={https://doi.org/10.1038/ng.2285},
+  DOI={10.1038/ng.2285},
   publisher={Nature Publishing Group}
 }
