### 2023-05-10 0961ed2
fix DOI issue detected by JOSS editorialbot
