+++ paper/paper.bib
@@ -6,7 +6,7 @@
   number={10},
   pages={1350--1360},
   year={2021},
-  DOI={https://doi.org/10.1038/s41559-021-01526-9},
+  DOI={10.1038/s41559-021-01526-9},
   publisher={Nature Publishing Group UK London}
 }
 
@@ -18,7 +18,7 @@
   number={1},
   pages={e1005703},
   year={2016},
-  DOI={https://doi.org/10.1371/journal.pgen.1005703},
+  DOI={10.1371/journal.pgen.1005703},
   publisher={Public Library of Science San Francisco, CA USA}
 }
 
