### 2023-05-10 #5455 @CheWeiChang1992
@Kevin-Mattheus-Moerman
Thank you very much for your assistance.
1. I've solved the invalid DOIs detected by the editorial bot.
I also checked references with `@editorialbot check references` and update PDF with `@editorialbot generate pdf` as you suggested.
2. Our MIT `LICENSE` is located at `./ggoutlier/LICENSE`, probably that is why the editorial bot could not find it. I've made a copy of `LICENSE` in the main directory.
Let me know if there is anything that should be fixed.
Thank you.
