### 2023-05-21 #5455 @Kevin-Mattheus-Moerman
Thanks @CheWeiChang1992 for making those changes. Note I have just flagged this submission for scope review by the editorial board. This is mainly to check if the software conforms to our [substantial scholarly effort criteria](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort) in terms of both size and functionality. This review should be completed within 2 weeks.
