### 2023-05-30 #5455 @Kevin-Mattheus-Moerman
@CheWeiChang1992

As part of the scope review one our editors pointed our the following:

> The GitHub repository is not set up in such a way that one can install it as one would normally do, because the package itself is hidden in the sub-directory named `ggoutlier`. In addition, it is not on CRAN, which is definitely a minus.

Can you please address these points before we proceed? Thanks.
