### 2023-05-30 9f5398d
remove redundant R example in get_GGNet_adjacency_matrix.R -> fix errors of R CMD check
