### 2023-05-30 bca5cb0
fix ERROR and WORNING for R CMD check --as-cran
