+++ ggoutlier/R/data.R
@@ -1,0 +1,14 @@
+#' Ancestry coefficients of IPK barley landraces
+#'
+#' @name ipk_anc_coef
+#' @docType data
+#' @format matrix
+NULL
+
+#' Geographical origins of IPK barley landraces
+#'
+#' @name ipk_geo_coord
+#' @docType data
+#' @format data.frame
+
+NULL
