+++ ggoutlier/man/ipk_anc_coef.Rd
@@ -1,0 +1,12 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/data.R
+\docType{data}
+\name{ipk_anc_coef}
+\alias{ipk_anc_coef}
+\title{Ancestry coefficients of IPK barley landraces}
+\format{
+matrix
+}
+\description{
+Ancestry coefficients of IPK barley landraces
+}

[…]
