### 2023-05-31 ecc09d2
update GGoutlieR package -> address issues identified by R CMD check --as-cran
