+++ R/zzz.R
@@ -1,8 +1,0 @@
-.onLoad = function(library, pkg)
-{
-  if(interactive())
-  {
-    packageStartupMessage("# Load GGoutlieR v1.0.0")
-  }
-
-}
