+++ man/ggoutlier_example.Rd
@@ -1,0 +1,12 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/data.R
+\docType{data}
+\name{ggoutlier_example}
+\alias{ggoutlier_example}
+\title{Example of `ggoutlier` output}
+\format{
+list
+}
+\description{
+Example of `ggoutlier` output
+}
