### 2023-05-31 fab1bf0
fix an error in parallel computation due to miss spelling of package name
