### 2023-06-01 e8f7316
fix the issues reported by R CRAN auto check -> abolish the progress bar (use doParallel instead of doSNOW); make figure output to be optional; reduce the sample size for running the example code of ggoutlier
