+++ R/find_optimalK_geneticKNN.R
@@ -13,20 +13,23 @@ find_optimalK_geneticKNN <- function(geo_coord,
 
   ## parallel computation (if `cpu` > 1)
   if(do_par){
-    doSNOW::registerDoSNOW(cl)
+    doParallel::registerDoParallel(cl)
     kindx <- seq(from = min(klim), to = max(klim), by = 1)
     parallel::clusterExport(cl = cl,
                             unclass(lsf.str(envir = asNamespace("GGoutlieR"),
                                             all = T)),
                             envir = as.environment(asNamespace("GGoutlieR"))
     )
-    ## setup a progress bar for foreach
-    pb <- txtProgressBar(max = max(klim), min = min(klim), style = 3)
-    progress <- function(n){setTxtProgressBar(pb, n)}
-    opts <- list(progress = progress)
-    clusterExport(cl, "opts", envir = environment())
 
-    all.D <- foreach(k = kindx, .packages=c('geosphere','FastKNN'), .combine="c", .options.snow = opts) %dopar% {
+    # NOTE: abolish progress bar to meet the requirement of CRAN submission
+    ## setup a progress bar for foreach
+    #pb <- txtProgressBar(max = max(klim), min = min(klim), style = 3)
+    #progress <- function(n){setTxtProgressBar(pb, n)}
+    #opts <- list(progress = progress)
+    #clusterExport(cl, "opts", envir = environment())
+    #all.D <- foreach(k = kindx, .packages=c('geosphere','FastKNN'), .combine="c", .options.snow = opts) %dopar% {
+
+    all.D <- foreach(k = kindx, .packages=c('geosphere','FastKNN'), .combine="c") %dopar% {
       knn.indx <- find_gen_knn(pgdM, k=k)
       # KNN prediction
       pred.geo_coord <- pred_geo_coord_knn(geo_coord = geo_coord,
