### 2023-06-01 51eccd0
fix the issue reported by R CRAN auto check -> modify MIT License
