### 2023-06-02 #5455 @CheWeiChang1992
> @CheWeiChang1992
>
> The scope review has concluded this work may be in scope for JOSS. However, as part of the scope review, the following was pointed out:
>
> > The GitHub repository is not set up in such a way that one can install it as one would normally do, because the package itself is hidden in the sub-directory named `ggoutlier`. In addition, it is not on CRAN, which is definitely a minus.
>
> Can you please address these points before we proceed? Thanks.

Thank you very much for pointing out the issues.
- I moved the package to the main directory, so now it is able to be installed with the `install_github` function in the R package `devtools`.
- I submitted the package to CRAN and it is now waiting for manual review by the CRAN team (see below). I'll update you if there is any news.
> Dear maintainer,
> package GGoutlieR_1.0.0.tar.gz has been auto-processed and is pending a manual inspection of this new CRAN submission. A CRAN team member will typically respond to you within the next 10 working days. For technical reasons you may receive a second copy of this message when a team member triggers a new check.
> Log dir: <https://win-builder.r-project.org/incoming_pretest/GGoutlieR_1.0.0_20230601_151744/>
> The files will be removed after roughly 7 days.
> Installation time in seconds: 7
> Check time in seconds: 93
> R Under development (unstable) (2023-05-31 r84480 ucrt)
