### 2023-06-05 #5455 @martinfleis
@CheWeiChang1992 if you have any suggestions for potential reviewers then please mention them here in this thread (without tagging them with an @). In addition, [this list of people](https://bit.ly/joss-reviewers) have already agreed to review for JOSS and may be suitable for this submission (please start at the bottom of the list). Genetics is a bit out of my comfort zone, so I'd appreciate some helping hand here :). Thanks!
