### 2023-06-07 6cfabb7
modify DESCRIPTION accrodign to the suggestion of CRAN team
