+++ DESCRIPTION
@@ -1,10 +1,10 @@
 Package: GGoutlieR
 Type: Package
-Title: An R Package Used to Identify Individuals with Unusual Geo-Genetic Patterns
+Title: Identify Individuals with Unusual Geo-Genetic Patterns
 Version: 1.0.0
 Author: Che-Wei Chang, Karl Schmid 
 Maintainer: Che-Wei Chang <cheweichang92@gmail.com>
-Description: Identifies individuals with unusual association patterns of genetics and geography. It produces a table of outliers with statistics and visualizes unusual geo-genetic patterns on a geographical map. This is useful for landscape genomics studies to discover samples that violate the isolation-by-distance assumption.
+Description: Identify and visualize individuals with unusual association patterns of genetics and geography. It detects potential outliers that violate the isolation-by-distance assumption using the K-nearest neighbor approach. You can obtain a table of outliers with statistics and visualize unusual geo-genetic patterns on a geographical map. This is useful for landscape genomics studies to discover individuals with unusual geography and genetics associations from a large biological sample.
 License: MIT + file LICENSE
 Depends:
     R (>= 3.5.0)
