### 2023-06-07 3916ae9
modify the author field in DESCRIPTION according to the suggestion of CRAN team
