+++ DESCRIPTION
@@ -3,7 +3,18 @@ Type: Package
 Title: Identify Individuals with Unusual Geo-Genetic Patterns
 Version: 1.0.0
 Author: Che-Wei Chang, Karl Schmid 
-Maintainer: Che-Wei Chang <cheweichang92@gmail.com>
+Authors@R:
+  c(person(given = "Che-Wei",
+           family = "Chang",
+           role = c("aut", "cre"),
+           email = "cheweichang92@gmail.com",
+           comment = c(ORCID = "0000-0001-6704-3732"),
+    person(given = "Karl",
+           family = "Schmid",
+           role = c("ths"),
+           email = "karl.schmid@uni-hohenheim.de",
+           comment = c(ORCID = "0000-0001-5129-895X"),
+           )
 Description: Identify and visualize individuals with unusual association patterns of genetics and geography. It detects potential outliers that violate the isolation-by-distance assumption using the K-nearest neighbor approach. You can obtain a table of outliers with statistics and visualize unusual geo-genetic patterns on a geographical map. This is useful for landscape genomics studies to discover individuals with unusual geography and genetics associations from a large biological sample.
 License: MIT + file LICENSE
 Depends:
