### 2023-06-07 9e044cc
fix an error in the author field of DESCRIPTION
