### 2023-06-07 b3599dc
make GGoutlieR to recover graphic parameters according to the suggestion of CRAN team
