+++ R/ggoutlier.R
@@ -47,6 +47,18 @@
 #'                              min_nn_dist = 1000,
 #'                              multi_stages = FALSE)
 #' summary_ggoutlier(mini_example)
+#'
+#' # DON'T RUN: this analysis will take a few minutes
+#' # full_example <- ggoutlier(geo_coord = ipk_geo_coord,
+#' #                           gen_coord = ipk_anc_coef,
+#' #                           klim = c(3,6),
+#' #                           p_thres = 0.01,
+#' #                           cpu = 2,
+#' #                           method = "composite",
+#' #                           verbose = FALSE,
+#' #                           min_nn_dist = 1000,
+#' #                           multi_stages = FALSE)
+#'
 #' @export
 
 ggoutlier <- function(geo_coord,
