### 2023-06-07 ca6ab9b
minor update of tutorial -> the default of the latest GGoutlieR will suppress the auto plotting feature according to the request of CRAN team -> no PDF generated by `ggoutlier()` if users do not set `make_fig = TRUE`
