### 2023-06-07 #5455 @CheWeiChang1992
> @CheWeiChang1992 if you have any suggestions for potential reviewers then please mention them here in this thread (without tagging them with an @). In addition, [this list of people](https://bit.ly/joss-reviewers) have already agreed to review for JOSS and may be suitable for this submission (please start at the bottom of the list). Genetics is a bit out of my comfort zone, so I'd appreciate some helping hand here :). Thanks!

Thank you for the information.
We would like to suggest the following scientists in the field of landscape genomics as potential reviewers:
- Geoffrey House and/or Matthew Hahn (github user name: hahnlab-user)
- Gideon Bradburd (github user name: gbradburd)
- Peter Ralph (github user name: petrelharp)

For the list of JOSS reviewers (https://bit.ly/joss-reviewers), it seems that I need access. I sent a request for the list through my private Google account (jerry122116@gmail.com). I'd be grateful if you could confirm my access request. I can then have a look at the list.
Thank you for your assistance!
