### 2023-06-08 2948178
modify the documentation according to the request of CRAN team
