### 2023-06-09 bfba246
update user manual and DESCRIPTION -> add DOI of GGoutlieR paper (biRxiv)
