+++ R/summary_ggoutlier.R
@@ -2,7 +2,7 @@
 #' @description
 #' Get a summary table from the `ggoutlier` output
 #' @param  ggoutlier_res output from the function `ggoutlier`
-#' @return a table contains IDs of outliers and p-values.
+#' @returns a table contains IDs of outliers and p-values.
 #' @examples
 #' library(GGoutlieR)
 #' data(ggoutlier_example) # get an example output of ggoutlier
