### 2023-06-12 #5455 @Winter
> > Sorry! The list of JOSS reviewers is now here https://reviewers.joss.theoj.org/reviewers. Can you check that one?
>
> Thank you for the JOSS reviewer list.
>
> I selected three people from the list who may be suitable as potential reviewers.
>
> * Tyler Chafin (github name: tkchafin)
> * Maximilian Hesselbarth (github name: mhesselbarth)
> * David Winter (github name: winter)
>
> By the way, now the package is available on the CRAN. Please let me know if you have any suggestions. Thank you!

Hey Che-Wei Chang!

You may want to get in contact with _David Winter_ to confirm his GitHub handle as I'm not David :)

Best Regards,
Edward (@Winter)
