### 2023-06-14 kjschmidlab/GGoutlieR#1 @rsbivand
Vulnerability to retiring spatial infrastructure packages
For background see: https://r-spatial.org/r/2023/05/15/evolution4.html and https://r-spatial.github.io/evolution/. This package imports `sp`, but this month `sp` will switch to relying on `sf` in place of `rgdal` to access upstream external software. When running forthcoming `sp` and without retiring packages on the library path, the CMD check log file is:
[00check.log](https://github.com/kjschmidlab/GGoutlieR/files/11744444/00check.log) with install failure
[00install.zip](https://github.com/kjschmidlab/GGoutlieR/files/11744449/00install.zip) cause by importing `rworldmap`, which imports `maptools`. Once `rworldmap` is updated https://github.com/AndySouth/rworldmap/issues/20, you may need to add `sf` to Suggests:.
