### 2023-06-19 #5455 @martinfleis
Hi @ tgerke, would you be able to review this submission to JOSS?

@mhesselbarth thanks for letting me know!
