### 2023-07-17 #5455 @martinfleis
@tkchafin thanks!

@btmartin721 would you be willing to review this package for JOSS?
