### 2023-07-24 #5687 @editorialbot
[REVIEW]: GGoutlieR: an R package to identify and visualize unusual geo-genetic patterns of biological samples
**Submitting author:** @CheWeiChang1992 (<a href="http://orcid.org/0000-0001-6704-3732">Che-Wei Chang</a>)
**Repository:** https://github.com/kjschmidlab/GGoutlieR
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.2
