### 2023-08-28 kjschmidlab/GGoutlieR#2 @tkchafin
[JOSS REVIEW] Add community guidelines docs
- Add instruction on how you would like users to seek support/ report problems in the README.Md (usually this is just linking to the Issues page)
- Add instructions on how to contribute
