### 2023-08-28 kjschmidlab/GGoutlieR#3 @tkchafin
[JOSS REVIEW] object k.sel.plot not found 
Received this error trying to run with the example data

```
library(GGoutlieR)

data('ipk_anc_coef')
data('ipk_geo_coord')

pthres=0.025

res <- ggoutlier(geo_coord = ipk_geo_coord,
                 gen_coord = ipk_anc_coef,
                 p_thres=pthres,
                 plot_dir="",
                 cpu=4,
                 klim=c(3,50),
                 method="composite",
                 verbose=T,
                 min_nn_dist=1000)
```

Output:
```
Parallelize computation using 4 cores
Find samples with identical geographical coordinates.
Ignore neighbors whose pairwise distance is less than `min_nn_dist` when searching KNNs.

Ignore neighbors within 10 unit(s) of distance (the default unit of distance is m; `min_nn_dist` is adjusted automatically according to `s`)
Find samples with zero genetic distances.
Add 10^-6 unit of distance to individual pairs

 `k_geneticKNN` is NULL; searching for optimal k between 3 and 50
this process can take time...
Error in paste("\n The optimal k_geneticKNN is ", opt.k, ". Its figure is saved at ",  :
  object 'k.sel.plot' not found
```

Running in R4.1.3 in RStudio on Windows. I got an alert via the firewall, so I'm not sure if this has something to do with one of the dependencies trying to download a base map or something? If that is the case, you may consider either shipping with a basic basemap, swapping dependencies, or catching the error and printing on a blank canvas -- in either case some sort of error handling would be beneficial. Not sure if this is related to issue https://github.com/kjschmidlab/GGoutlieR/issues/1 or not
