### 2023-08-28 kjschmidlab/GGoutlieR#4 @tkchafin
[JOSS REVIEW] Minor suggestions 
Some minor suggestions:
- Link the CRAN docs on README, making sure the reference manual pdf file is easily findable on GitHub
- Might want to remove the redundant LICENSE file? Doesn't seem to add anything beyond the LICENSE.Md?
- Would be nice to put the simple example from the manuscript into the README.Md, so users can quickly test if their installation is working
- Wasn't able to install on ARM64 due to dependencies either being unavailable or failing to compile. This might suggest the ecosystem of dependencies is not the most robust -- you might consider (in the long term) working to reduce dependencies, or seek dependencies which will be more reliable across systems
- MS 'Algorithm' section: Description of D not very clear; isn't this just MSE? In that case it would be more straightforward to describe it as such. Also, it would be good to highlight here that the method can be used either to define geographical outliers with respect to genetic ‘neighbors’, or genetic outliers with respect to geographical ‘neighbors’, as the potential use cases for each would differ, so you want to make sure users are aware that each can be done using your package
