### 2023-08-30 kjschmidlab/GGoutlieR#3 @tkchafin
```
> library(GGoutlieR)
>
> data('ipk_anc_coef')
> data('ipk_geo_coord')
>
> pthres=0.025
>
> res <- ggoutlier(geo_coord = ipk_geo_coord,
+                  gen_coord = ipk_anc_coef,
+                  p_thres=pthres,
+                  plot_dir="",
+                  cpu=4,
+                  klim=c(3,50),
+                  method="composite",
+                  verbose=T,
+                  min_nn_dist=1000)

 Parallelize computation using 4 cores
Find samples with identical geographical coordinates.
Ignore neighbors whose pairwise distance is less than `min_nn_dist` when searching KNNs.

Ignore neighbors within 10 unit(s) of distance (the default unit of distance is m; `min_nn_dist` is adjusted automatically according to `s`)
Find samples with zero genetic distances.
Add 10^-6 unit of distance to individual pairs

 `k_geneticKNN` is NULL; searching for optimal k between 3 and 50
this process can take time...
Error in ggoutlier_compositeKNN(geo_coord = geo_coord, gen_coord = gen_coord,  :
  object 'k.sel.plot' not found
```
