### 2023-08-30 #5687 @tkchafin
I've gone as far as I can in the review process for now as I've been unable to successfully run or install the package across 3 machines (MacOS ARM, and 2 Windows). This issue was logged in https://github.com/kjschmidlab/GGoutlieR/issues/3

I don't have time to dig into the code and track down the source of this this week, but it may be related to the package relying on an apparently unstable set of dependencies, and relies on a number which will be retired in October of this year:
```
The legacy packages maptools, rgdal, and rgeos, underpinning the sp package,
which was just loaded, will retire in October 2023.
Please refer to R-spatial evolution reports for details, especially
https://r-spatial.org/r/2023/05/15/evolution4.html.
It may be desirable to make the sf package available;
package maintainers should consider adding sf to Suggests:.
The sp package is now running under evolution status 2
     (status 2 uses the sf package in place of rgdal)
Please note that 'maptools' will be retired during October 2023,
plan transition at your earliest convenience (see
https://r-spatial.org/r/2023/05/15/evolution4.html and earlier blogs
for guidance);some functionality will be moved to 'sp'.
```
This was also pointed out by another user in https://github.com/kjschmidlab/GGoutlieR/issues/1

For now, I've also suggested the creation of contribution guidelines (https://github.com/kjschmidlab/GGoutlieR/issues/2), and some various minor suggestions (https://github.com/kjschmidlab/GGoutlieR/issues/4) which I don't consider strictly necessary but hopefully some are helpful.

I have no issues with the manuscript or methodological approach, and congratulate the authors for developing a useful package that addresses a common issue in landscape genetics, which I look forward to using in my own research once these minor issues are ironed out. Will resume functionality testing after the authors have had a chance to take a look at the various issues.
