### 2023-09-06 f73f12c
fix the error regarding missing k.sel.plot
