+++ R/detect_outlier_composite.R
@@ -209,7 +209,12 @@ ggoutlier_compositeKNN <- function(geo_coord,
     }
 
     k_geneticKNN = opt.k
-    if(verbose) cat(paste("\n The optimal k_geneticKNN is ",opt.k,". Its figure is saved at ", k.sel.plot," \n", sep = ""))
+    if(make_fig){
+      if(verbose) cat(paste0("\n The optimal k_geneticKNN is ",opt.k,". Its figure is saved at ", k.sel.plot," \n"))
+    }else{
+      if(verbose) cat(paste0("\n The optimal k_geneticKNN is ",opt.k," \n"))
+    }
+
 
   }
   if(is.null(k_geneticKNN)){stop("k_geneticKNN is NULL!")}
