### 2023-09-27 abc5957
Issue#1: retireing sp package -> change R function for geographical distance calculation -> use sf package
