### 2023-09-27 b7780e1
Issue#1: retiring sp package -> abort the original plot function depending on rworldmap -> use rnaturalearth and sf instead (the new function was finished but not tested yet)
