+++ R/plot_GGoutlieR.R
@@ -14,7 +14,7 @@
 #' @param  p_thres numeric. A value of significant level. Only outliers (p values less than `p_thres`) are mapped on a geographical map if `p_thres` is provided (the default is `NULL`). This argument is optional.
 #' @param  color_res integer. The resolution of color scale.
 #' @param  dot_cex numeric. The size of dots denoting the positions of samples on a geographical map.
-#' @param  map_type string. The type of plot to draw. It can be `"geographic_knn"`, `"genetic_knn"` and `"both"`.
+#' @param  map_type string. The type of plot to draw. It can be `"geographic_knn"`, `"genetic_knn"` and `"add_both"`.
 #' @param  select_xlim vector. Values controlling longitude boundaries of a window to select outliers to present on a geographical map. The default is `select_xlim = c(-180,180)`.
 #' @param  select_ylim vector. Values controlling latitude boundaries of a window to select outliers to present on a geographical map. The default is `select_ylim = c(-90,90)`.
 #' @param  plot_xlim vector. Values controlling longitude boundaries of a map.
@@ -22,14 +22,15 @@
 #' @param  only_edges_in_xylim logic. only the edges with starting points within the given `select_xlim` and `select_ylim` will display on a geographical map. If `FALSE`, the edges out of the given boundaries will be removed from your plot. The default is `TRUE`.
 #' @param  pie_r_scale numeric. A scale controlling the radius of pie charts
 #' @param  red_alpha numeric. A value controlling the transparency of red lines. the default is 0.8
-#' @param  map_resolution string. The resolution of the geographical map. See details in the manual of `rworldmap::getMap()`
+#' @param  map_resolution a character string. The resolution of the geographical map. See details of the `scale` argument in the manual of `rnaturalearth::ne_countries()`. The default is `map_resolution = "small"`
 #' @param  show_knn_pie logic. If `TRUE`, the ancestry coefficients of K nearest neighbors of significant samples will display on the map. The default is `FALSE`.
-#' @param  show_col_bar logic. If `TRUE`, a color key will be added to the output graph.
+#' @param  show_col_legend logic. If `TRUE`, a color key will be added to the output graph.
 #' @param  which_sample a string vector of sample ID(s). If users want to only show specific sample(s)
 #' @param  add_benchmark_graph logic. If `TRUE`, a benchmark graph with only pie charts of ancestry coefficients for comparison with the outlier graph.
 #' @param  vertical_plots logic. If `TRUE`, a benchmark graph and outlier graph will be combined in a vertical direction.
 #' @param  adjust_p_value_projection logic. If `TRUE`, the function will perform KNN prediction by forcing K=1 and compute new p-values for visualization.
-#' @returns a geographical map with colored lines showing sample pairs with unusual geo-genetic associations.
+#' @param  linewidth_range numeric. A vector of two values. It is used to control the minimal and maximal width of KNN network on the geographical map.
+#' @returns a list of ggplot object(s). The plot is geographical map with colored lines showing sample pairs with unusual geo-genetic associations.
 #' @examples
 #' library(GGoutlieR)
[…]
