### 2023-09-27 1dfa91f
fix an error caused by object class -> convert geographical distances from Units object to vector with `as.vector`
