### 2023-09-28 b962538
fix an error due to geo.dM class -> reform st_distance output to be a matrix
