### 2023-09-28 e97d2a2
minor change in description of plot_ggoutlier
