+++ NAMESPACE
@@ -15,9 +15,7 @@ importFrom("stats", "cor", "dgamma", "dist", "pgamma", "qgamma",
 importFrom("utils", "lsf.str", "setTxtProgressBar", "txtProgressBar")
 importFrom("parallel", "detectCores","clusterExport", "makeCluster", "stopCluster")
 importFrom("foreach", "%dopar%", "foreach")
-importFrom("mapplots", "add.pie")
-importFrom("rworldmap", "getMap")
-importFrom("plotrix", "color.legend")
 importFrom("scales","alpha")
 importFrom("stats4","mle")
 importFrom("RColorBrewer", "brewer.pal", "brewer.pal.info")
+importFrom("sf", "st_distance", "st_as_sf")
