+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Package: GGoutlieR
 Type: Package
 Title: Identify Individuals with Unusual Geo-Genetic Patterns
-Version: 1.0.0
+Version: 1.0.1
 Authors@R:
   c(person(given = "Che-Wei",
            family = "Chang",
@@ -19,22 +19,26 @@ License: MIT + file LICENSE
 Depends:
     R (>= 3.5.0)
 Imports:
-    geosphere,
     stats4, 
     FastKNN,
     foreach,
     doParallel,
     iterators,
     parallel,
-    rworldmap,
     scales,
-    plotrix,
-    mapplots, 
     RColorBrewer,
-    rworldxtra,
-    dichromat,
-    sp
+    ggforce, 
+    rlang, 
+    ggfun, 
+    stats, 
+    tidyr, 
+    dplyr, 
+    utils,
+    rnaturalearth,
+    rnaturalearthdata, 
+    sf,
+    ggplot2,
+    cowplot
 Encoding: UTF-8
 LazyData: true
-RoxygenNote: 7.1.1
-
+RoxygenNote: 7.2.3
