### 2023-09-28 11b978d
remove redundant code about color alpha to prevent errors
