### 2023-09-29 a6461a3
fix bug in plot_ggoutlier
