+++ R/plot_GGoutlieR.R
@@ -134,6 +134,12 @@ plot_ggoutlier <- function(ggoutlier_res,
     if(is.null(dot_cex)){
       dot_cex <- min(linewidth_range)
     }
+    if(is.null(anc_coef) & is.null(gen_coord)){
+      stop("please provide at least either `anc_coef` or `gen_coord`")
+    }else{
+      if(is.null(gen_coord)){gen_coord <- anc_coef}
+      if(is.null(anc_coef)){anc_coef <- gen_coord}
+    }
     # assign colors if pie_color is NULL
     if(!is.null(anc_coef) & is.null(pie_color)){
       # check if sample sizes match
@@ -155,12 +161,7 @@ plot_ggoutlier <- function(ggoutlier_res,
     warning("`p_thres` is `NULL`. ancestry coefficients will NOT be projected to your geographical map. the pie charts of outliers could be added to the map if you give `p_thres`.")
   }
 
-  if(is.null(anc_coef) & is.null(gen_coord)){
-    stop("please provide at least either `anc_coef` or `gen_coord`")
-  }else{
-    if(is.null(gen_coord)){gen_coord <- anc_coef}
-    if(is.null(anc_coef)){anc_coef <- gen_coord}
-  }
+
 
   ## geographical data: data.frame to sf format
   geo_data_df <- data.frame(ID = rownames(geo_coord),
