### 2023-09-29 5ef3ea2
fix bug in plot_ggoutlier
