### 2023-09-29 63e39bc
adjust the output plot position of plot_ggoutlier
