### 2023-10-01 a99d01a
address Issue#1 : get rid of geosphere package which replies on sp
