+++ R/Dgeo_geneticKNN.R
@@ -27,6 +27,8 @@ pred_geo_coord_knn <- function(geo_coord, pgdM, knn.indx, w_power){
 
 ## a function to calculate Dg
 cal_Dgeo <- function(pred.geo_coord, geo_coord, scalar){
-  geodist <- (sapply(1:nrow(geo_coord),function(a){geosphere::distm(x = geo_coord[a,], y = pred.geo_coord[a,])/scalar}))
+  geo_coord_sf <- sf::st_as_sf(geo_coord, coords = c("x", "y"), crs = 4326)
+  pred.geo_coord_sf <- sf::st_as_sf(pred.geo_coord, coords = c("x", "y"), crs = 4326)
+  geodist <- (sapply(1:nrow(geo_coord),function(a){as.vector(sf::st_distance(x = geo_coord_sf[a,], y = pred.geo_coord_sf[a,]))/scalar}))
   return(geodist)
 } # cal_Dgeo end
