### 2023-10-01 5e2dd94
name the column of `pred_geo_coord_knn` output to meet the format requirement of st_as_sf function
