### 2023-10-01 70adc03
make `pred_geo_coord_knn` output to be data.frame to meet the format requirement of st_as_sf function
