### 2023-10-02 #5687 @martinfleis
Hi @CheWeiChang1992, I see that you did not respond to any of the comments made by @tkchafin here or in the original repository. The JOSS is a bit different that the other journal in that it assumes a _conversation_ between authors and reviewers. Would it be possible to let us know when you'll be able to get back to the submission?
