### 2023-10-02 #5687 @CheWeiChang1992
Hi @martinfleis thank you for your kind reminder. I am still working on the replacement of the `sp` package and the other dependencies using the `sp` package. Because of these changes, I will update the figures in the manuscript and tutorial (the new version of `GGoutlieR` will use `ggplot2` and `rnaturalearth` packages to make geographical plots to replace `rworldmap`). I will work on the other comments after updating the functions of the `GGoutlieR` package.
