### 2023-10-03 ae6cfe2
update `cal_Dgeo` function -> reduce computational time
