### 2023-10-03 060140e
add print out message to show the progress of multi-stage test
