### 2023-10-04 d207549
fix warning from ggplot -> fix incorrect input data format
