### 2023-10-04 d011123
add checkpoint for plot_dir to ggoutlier
