### 2023-10-08 0caa69e
remove redundant codes; adjust output plot
