+++ R/plot_GGoutlieR.R
@@ -21,18 +21,16 @@
 #' @param  plot_ylim vector. Values controlling latitude boundaries of a map.
 #' @param  only_edges_in_xylim logic. only the edges with starting points within the given `select_xlim` and `select_ylim` will display on a geographical map. If `FALSE`, the edges out of the given boundaries will be removed from your plot. The default is `TRUE`.
 #' @param  pie_r_scale numeric. A scale controlling the radius of pie charts
-#' @param  red_alpha numeric. A value controlling the transparency of red lines. the default is 0.8
-#' @param  map_resolution a character string. The resolution of the geographical map. See details of the `scale` argument in the manual of `rnaturalearth::ne_countries()`. The default is `map_resolution = "small"`
+#' @param  map_resolution a character string. The resolution of the geographical map. See details of the `scale` argument in the manual of `rnaturalearth::ne_countries()`. The default is `map_resolution = "medium"`
 #' @param  show_knn_pie logic. If `TRUE`, the ancestry coefficients of K nearest neighbors of significant samples will display on the map. The default is `FALSE`.
 #' @param  show_col_legend logic. If `TRUE`, a color key will be added to the output graph.
 #' @param  which_sample a string vector of sample ID(s). If users want to only show specific sample(s)
 #' @param  add_benchmark_graph logic. If `TRUE`, a benchmark graph with only pie charts of ancestry coefficients for comparison with the outlier graph.
-#' @param  vertical_plots logic. If `TRUE`, a benchmark graph and outlier graph will be combined in a vertical direction.
 #' @param  adjust_p_value_projection logic. If `TRUE`, the function will perform KNN prediction by forcing K=1 and compute new p-values for visualization.
 #' @param  linewidth_range numeric. A vector of two values. It is used to control the minimal and maximal width of KNN network on the geographical map.
 #' @param  plot_labels character. A string of labels for plots. The default is `plot_labels = "auto"`. This parameter is for `cowplot::plot_grid`.
 #'
-#' @returns ggplot object. The plot is geographical map with colored lines showing sample pairs with unusual geo-genetic associations.
+#' @returns ggplot object. The plot is geographical map(s) with colored lines showing sample pairs with unusual geo-genetic associations.
 #' @examples
 #' library(GGoutlieR)
 #' data("ipk_anc_coef") # get ancestry coefficients
@@ -47,7 +45,7 @@
 #'                select_ylim = c(10,62),
 #'                plot_xlim = c(-20,140),
 #'                plot_ylim = c(10,62),
-#'                pie_r_scale = 1.2,
+#'                pie_r_scale = 1.8,
 #'                map_resolution = "medium")
 #'
 #' @export
@@ -72,13 +70,11 @@ plot_ggoutlier <- function(ggoutlier_res,
                            plot_ylim = NULL,
                            only_edges_in_xylim = TRUE,
                            pie_r_scale = 1,
-                           #red_alpha = 0.8,
-                           map_resolution = "small",
+                           map_resolution = "medium",
[…]
