+++ man/plot_ggoutlier.Rd
@@ -21,13 +21,14 @@ plot_ggoutlier(
   plot_ylim = NULL,
   only_edges_in_xylim = TRUE,
   pie_r_scale = 1,
-  map_resolution = "small",
+  map_resolution = "medium",
   show_knn_pie = FALSE,
   show_col_bar = TRUE,
   which_sample = NULL,
   add_benchmark_graph = TRUE,
   adjust_p_value_projection = FALSE,
-  linewidth_range = c(0.5, 3)
+  linewidth_range = c(0.5, 3),
+  plot_labels = "auto"
 )
 }
 \arguments{
@@ -63,7 +64,7 @@ plot_ggoutlier(
 
 \item{pie_r_scale}{numeric. A scale controlling the radius of pie charts}
 
-\item{map_resolution}{a character string. The resolution of the geographical map. See details of the `scale` argument in the manual of `rnaturalearth::ne_countries()`. The default is `map_resolution = "small"`}
+\item{map_resolution}{a character string. The resolution of the geographical map. See details of the `scale` argument in the manual of `rnaturalearth::ne_countries()`. The default is `map_resolution = "medium"`}
 
 \item{show_knn_pie}{logic. If `TRUE`, the ancestry coefficients of K nearest neighbors of significant samples will display on the map. The default is `FALSE`.}
 
@@ -75,14 +76,12 @@ plot_ggoutlier(
 
 \item{linewidth_range}{numeric. A vector of two values. It is used to control the minimal and maximal width of KNN network on the geographical map.}
 
-\item{red_alpha}{numeric. A value controlling the transparency of red lines. the default is 0.8}
+\item{plot_labels}{character. A string of labels for plots. The default is `plot_labels = "auto"`. This parameter is for `cowplot::plot_grid`.}
 
 \item{show_col_legend}{logic. If `TRUE`, a color key will be added to the output graph.}
-
-\item{vertical_plots}{logic. If `TRUE`, a benchmark graph and outlier graph will be combined in a vertical direction.}
 }
 \value{
-ggplot object. The plot is geographical map with colored lines showing sample pairs with unusual geo-genetic associations.
+ggplot object. The plot is geographical map(s) with colored lines showing sample pairs with unusual geo-genetic associations.
 }
 \description{
 Visualize geo-genetic patterns of outliers with their K nearest neighbors
@@ -107,7 +106,7 @@ plot_ggoutlier(ggoutlier_res = ggoutlier_example,
                select_ylim = c(10,62),
                plot_xlim = c(-20,140),
                plot_ylim = c(10,62),
-               pie_r_scale = 1.2,
+               pie_r_scale = 1.8,
                map_resolution = "medium")
 
 }
