### 2023-10-09 3366b51
address warnings and notes of R CMD check for CRAN
