+++ DESCRIPTION
@@ -23,22 +23,20 @@ Imports:
     FastKNN,
     foreach,
     doParallel,
-    iterators,
     parallel,
     scales,
     RColorBrewer,
     ggforce, 
     rlang, 
-    ggfun, 
     stats, 
     tidyr, 
-    dplyr, 
     utils,
     rnaturalearth,
-    rnaturalearthdata, 
     sf,
     ggplot2,
     cowplot
+Suggests:
+    rnaturalearthdata
 Encoding: UTF-8
 LazyData: true
 RoxygenNote: 7.2.3
