### 2023-10-09 eb900fd
update p values after replace geosphere with sf package for geographical distance calculation
