### 2023-10-09 f3df427
minor update of example codes in tutorial
