### 2023-10-09 38f1b1a
correct typo in help manual
