### 2023-10-09 kjschmidlab/GGoutlieR#1 @CheWeiChang1992
The calculation of geographical distance with `geosphere` is now replaced with `sf`.
`rworldmap` used in the plotting function is now replaced with `rnaturalearth`.
