### 2023-10-10 kjschmidlab/GGoutlieR#1 @rsbivand
@CheWeiChang1992 CMD check output without retiring packages on library path:
[00check.log](https://github.com/kjschmidlab/GGoutlieR/files/12854247/00check.log)

Quite a lot of undefined global functions or variables:
```
  aes coord_sf element_rect element_text geom_point geom_segment
  guide_legend guides labs logp scale_fill_manual scale_linewidth theme
  type value x x0 x1 y y0 y1
```
They look like missing import declarations possibly from `ggplot2`. Otherwise good to go, I think.
