### 2023-10-11 811a06e
minor update tutorial Rmd
