### 2023-10-11 b32bd8d
minor update figs and paper.md
