### 2023-10-12 34e7fa4
update README -> add simple example; add link to ref manual; add instruction of contributing
