+++ README.md
@@ -4,9 +4,13 @@
 It provides a summary table with heuristic p-values allowing users to easily identify outliers from thousands of biological samples.
 Moreover, its visualization tool enables users to present unusual geo-genetic association patterns on a geographical map.
 
+#### Details about GGoutlieR framework
+
+To find the detailed algorithm of `GGoutlieR`, please check the supplementary material of our manuscript [HERE](https://github.com/kjschmidlab/GGoutlieR/blob/master/paper/suppinfo.pdf) (`./paper/suppinfo.pdf`).
+
 ## Installation
 
-To install `GGoutlieR` v1.0.0, please use the R commands below:
+You can install `GGoutlieR` v1.0.1 either from CRAN or from this Github repository with the R commands below.
 
 ```
 ##-----------------------
@@ -19,21 +23,80 @@ install.packages("GGoutlieR")
 ##-----------------------
 
 ## install dependencies
-install.packages(c("geosphere","stats4","FastKNN","foreach","doParallel","iterators","parallel","rworldmap","scales","plotrix","mapplots", "RColorBrewer","rworldxtra","dichromat","sp"))
+install.packages(c("stats4","FastKNN","foreach","doParallel","parallel","scales", "RColorBrewer","ggforce", "rlang", "tidyr", "utils", "rnaturalearth", "rnaturalearthdata", "sf", "ggplot2","cowplot"))
 
-## install GGoutlieR
+## install GGoutlieR from Github
 library(devtools)
 install_git("https://github.com/kjschmidlab/GGoutlieR.git")
 ```
 
-## Tutorial
+## Get started
 
-You can find an example of `GGoutlieR` analysis using a global barley landrace collection [HERE](https://github.com/kjschmidlab/GGoutlieR/blob/master/vignettes/outlier_detection.pdf) (`./vignettes/outlier_detection.pdf`). 
+You can find a more detailed tutorial with an example of `GGoutlieR` using a global barley landrace collection [HERE](https://github.com/kjschmidlab/GGoutlieR/blob/master/vignettes/outlier_detection.pdf) (`./vignettes/outlier_detection.pdf`). 
 The VCF file of the barley landrace collection is available [HERE](https://github.com/kjschmidlab/GGoutlieR/blob/master/supp_data/IPKbarley_GBS1661landraces_LDpruned.vcf.gz) (`./supp_data/IPKbarley_GBS1661landraces_LDpruned.vcf.gz`).
 
-## Details about GGoutlieR framework
+### Simple example
 
-To find the detailed algorithm of `GGoutlieR`, please check the supplementary material of our manuscript [HERE](https://github.com/kjschmidlab/GGoutlieR/blob/master/paper/suppinfo.pdf) (`./paper/suppinfo.pdf`).
+```
+# example of barley landrace data 
+library(GGoutlieR)
+data("ipk_anc_coef") # get ancestry coefficients
+data("ipk_geo_coord") # get geographical coordinates
+
+# run ggoutlier (takes about 40 sec)
+ggoutlier_example <- 
+                ggoutlier(geo_coord = ipk_geo_coord,
+                          gen_coord = ipk_anc_coef,
+                          make_fig = FALSE,
+                          p_thres = 0.005, 
+                          cpu = 4, 
+                          klim = c(3,15), 
+                          method = "composite",
+                          verbose = FALSE,
+                          multi_stages = FALSE)) # switch off multi-stage test to reduce computational time
+                          
+# check the first few rows of the summary table
+head(summary_ggoutlier(ggoutlier_example))     
+
[…]
