### 2023-10-15 631ae9f
fix an error for ggoutlier composite method -> missing object `k.sel.plot` if setting `verbose=TRUE` and `make_fig=FALSE`
