### 2023-10-17 kjschmidlab/GGoutlieR#3 @CheWeiChang1992
Hello @tkchafin,
Thank you a lot for pointing out the issue.
The error is fixed in the v1.0.2, which is now available on CRAN.
