### 2023-10-18 f22d506
Minor grammatical corrections of paper.md
