### 2023-10-18 df1c0f7
Removed reference to specific version and minor change to text of README.md
