### 2023-10-18 a3d8da3
Update paper.md
add CRAN link to the Availability section; try to adjust size of fig 1
