### 2023-10-18 kjschmidlab/GGoutlieR#4 @CheWeiChang1992
Hello @tkchafin,
We appreciate your comments and suggestions for our package and manuscript. Please find our responses below.

> Some minor suggestions:
>
> * Link the CRAN docs on README, making sure the reference manual pdf file is easily findable on GitHub

 Thank you for your suggestion. We added the link to the "Reference manual" section of the README.md.

> * Might want to remove the redundant LICENSE file? Doesn't seem to add anything beyond the LICENSE.Md?

The LICENSE is required for the automatic check of the CRAN repository. So we think it would be better to keep it even though there is a MIT license.

> * Would be nice to put the simple example from the manuscript into the README.Md, so users can quickly test if their installation is working

We appreciate this nice suggestion. Indeed a simple example in README.md could be helpful for users. We added a section of "Simple example" to README.md

> * Wasn't able to install on ARM64 due to dependencies either being unavailable or failing to compile. This might suggest the ecosystem of dependencies is not the most robust -- you might consider (in the long term) working to reduce dependencies, or seek dependencies which will be more reliable across systems

Thank you for pointing out this issue. We didn't notice this as it passed the CRAN auto check. We will try to improve the package in this regard in the future version.

> * MS 'Algorithm' section: Description of D not very clear; isn't this just MSE? In that case it would be more straightforward to describe it as such. Also, it would be good to highlight here that the method can be used either to define geographical outliers with respect to genetic ‘neighbors’, or genetic outliers with respect to geographical ‘neighbors’, as the potential use cases for each would differ, so you want to make sure users are aware that each can be done using your package

Thank you for the question and suggestion. In our package, we defined a genetics-based D and a geography-based D. The former is, as you mentioned, equivalent to MSE, but the later one is not. To ensure the manuscript more focus on the implementation, we decided to describe details of our algorithm in the supplementary material. We updated the summary section of the manuscript to highlight the different respects of usage of our package.
