### 2023-10-18 kjschmidlab/GGoutlieR#2 @CheWeiChang1992
> * Add instruction on how you would like users to seek support/ report problems in the README.Md (usually this is just linking to the Issues page)
> * Add instructions on how to contribute

Thank you for the suggestion. We included the instruction of "Contributing" to the README.md.
