### 2023-10-18 kjschmidlab/GGoutlieR#3 @tkchafin
Latest version (1.0.2) fails to install (tried both from CRAN and github/devtools):
```
Error: object ‘scale_linewidth’ is not exported by 'namespace:ggplot2'
```
