### 2023-10-18 kjschmidlab/GGoutlieR#3 @tkchafin
Fixed it by updating ggplot2, might be worth tracking down when the change relevant to scale_linewidth was made and setting a version dependency
