### 2023-10-18 d2f915c
Fixed small typo in README example code
