### 2023-10-18 kjschmidlab/GGoutlieR#3 @tkchafin
Confirming that I can now run the code, so will close this as fixed. You may want to make your dependencies stricter (also noticed rnaturalearthdata isn't being installed as a dependency), but will go ahead and finish up my review over on JOSS. Good luck and thanks for fixing all of these issues :)
