### 2023-10-18 #5687 @tkchafin
Confirming that all the functional issues I encountered have been addressed with the most recent update (now up on CRAN), as well as minor comments, so I've completed my check list and have no further concerns. Best regards @CheWeiChang1992 and I look forward to using this package in my research!
