### 2023-10-18 3b87387
Update README.md
add more description about the package; remove the redundant right parenthesis in the example code
