### 2023-10-18 #5687 @CheWeiChang1992
Hi @tkchafin, thank you a lot for your review and patiently waiting. We addressed the open issues that you posted at [https://github.com/kjschmidlab/GGoutlieR/issues](https://github.com/kjschmidlab/GGoutlieR/issues).
Just let you know that there is a new release `GGoutlieR v1.0.2` available on CRAN. In this version:
* The error `object k.sel.plot not found` when setting `verbose = TRUE` is fixed
* The `sp` package and related dependencies are replaced with `sf` and other dependencies.
