### 2023-10-25 #5687 @btmartin721
The software functions as expected, it installed fine for me with no issues, the documentation is more than sufficient, and the software serves a need in the landscape genomics field. The plots are great! Very informative. They can serve to both identify outliers that may have been translocated, or who may be genetically similar for some other reason, such as temporal shifts in geographic features. I especially like the ability to detect both geographic and genetic outliers.

I will absolutely use this package in my ongoing research projects. I think it's a really cool paper and reads well. I have no concerns, and I wish @CheWeiChang1992 the best.

Kind regards,

-Bradley T. Martin
