### 2023-10-25 #5687 @CheWeiChang1992
Hello @martinfleis thank you for your kind help!
* the latest version of the `GGoutlieR` package is `v 1.0.2` which is available on the CRAN repository (https://cran.r-project.org/web/packages/GGoutlieR/index.html) and the GitHub repository.
* The latest release (v1.0.2) is archived on figshare with doi:10.6084/m9.figshare.24438388
