### 2023-10-30 346b349
fix spelling issue -> use American English only
