### 2023-10-30 #5687 @CheWeiChang1992
Hello @Kevin-Mattheus-Moerman thank you a lot for your help.
Regarding the points you mentioned:
* The version of FigShare archive is not related to the package version. I accidentally published duplicates when I edited the information and license of the FigShare archive. Its content is identical to the current version on master branch of our Github repository (`v1.0.2`) and CRAN repository (`v1.0.2`).
* I updated the manuscript on the Github repos.
Please let me know if any further issue needs to be addressed.
Thank you :)
