### 2023-11-01 #5687 @Kevin-Mattheus-Moerman
@CheWeiChang1992  thanks for making those changes. I think we are nearly there. I did notice the following final issue with the paper, can you have a look? :
- [ ] This sentence starts with `"`, but that is probably by mistake: `“Assuming isolation by distance, the geographical origins...`
