### 2023-11-01 ec07380
Update paper.md
fix typo
